Rename report library
